feat(core/cli/ipc): add transport-aware ArtAddress flow and 15-bit handling
Improve virtual-node and poll/write plumbing so ArtAddress writes can target bind transport endpoints and correctly program full 15-bit Net/Sub/Universe where supported.

Made-with: Cursor

+++ crates/lumenflow_cli/src/commands/send_all_packets.rs
@@ -68,7 +68,7 @@ pub async fn run(target: &str) -> Result<()> {
         [0; 4],
         [0; 4],
         0,
-        ArtAddressCommand::AcNone,
+        ArtAddressCommand::AcNone as u8,
     );
     socket.send_to(&pkt, target_addr).await?;
     count += 1;
